+++ src/components/Dish/Dish.tsx
@@ -4,34 +4,40 @@ import styled from 'styled-components'
 import { flexCenter } from '../../styles/mixins';
 import { DISHES, DISH_THEMES, IDishTheme } from '../../data/dishes';
 import DraggableEntity from '../DraggableEntity/DraggableEntity';
+import React from 'react';
+import List from '../List/List';
 
 export interface IDishProps {
   type: DISHES,
+  id: number
 }
 
-const Dish : FC<IDishProps> = ({type}) => {
+const Dish : FC<IDishProps> = React.memo(({type, id}) => {
   const dishTheme : IDishTheme = DISH_THEMES.get(type)!;
   return (
     <DraggableEntity>
       <StyledDishWrapper>
-        <Title>{type}</Title>
+        <Title>{type} {id}</Title>
         <DishBackground backImg={dishTheme.image} color={dishTheme.textColor}>
         </DishBackground>
       </StyledDishWrapper>
     </DraggableEntity>
   )
-}
+});
 
 const StyledDishWrapper = styled(motion.div)`
-    --dish-size: 20rem;
+  --dish-size: 22rem;
 
-    min-width: var(--dish-size);
-    height: var(--dish-size);
-    cursor: grab;
+  width: var(--dish-size);
+  height: var(--dish-size);
+  display: flex;
+  flex-direction: column;
 
-    &:active {
-      cursor: grabbing;
-    }
+  cursor: grab;
+
+  &:active {
+    cursor: grabbing;
+  }
 `;
 
 const Title = styled.p`
@@ -42,7 +48,7 @@ const Title = styled.p`
 
 const DishBackground = styled.figure<{backImg: string, color: string}>`
   width: 100%;
-  height: 100%;
+  flex: 1;
   background: url(${(props)=>props.backImg}) center/contain no-repeat;
   color: ${(props) => props.color};