+++ src/styles/globalStyles.tsx
@@ -1,8 +1,6 @@
 import { createGlobalStyle } from "styled-components";
 import { reset } from "./reset";
 import { BREAKPOINTS, variables } from "./variables";
-import backOpace from '../assets/back_opace.png';
-import logo from '../assets/logo.svg';
 
 const GlobalStyles = createGlobalStyle`
     ${reset}
@@ -16,21 +14,20 @@ const GlobalStyles = createGlobalStyle`
         }
 
         @font-face {
-            font-family: 'Kanit';
-            src: url('/fonts/kanit.ttf');
+            font-family: 'Minecraft';
+            src: url('/fonts/Minecraft.woff2') format('woff2'),
+                 url('/fonts/Minecraft.woff') format('woff'),
+                 url('/fonts/Minecraft.ttf') format('truetype');
         }
     }
 
     body {
         overflow: hidden;
         min-height: 100vh;
+        background-color: var(--color-console);
 
         font-size: var(--fs-body);
-        font-family: 'Kanit';
-        background: 
-            url(${backOpace}) bottom left / 30% no-repeat,
-            url(${logo}) top 1rem right 1rem no-repeat,
-            var(--gradient-background);
+        font-family: 'Minecraft';
     }
 `;